# LSM-based aggregation