--- a/changelogs/head.asciidoc
+++ b/changelogs/head.asciidoc
@@ -24,6 +24,6 @@ https://github.com/elastic/apm-server/compare/8.9\...main[View commits]
 
 [float]
 ==== Aggregation improvements
-- Replace aggregation implementation {pull}11117[11117]
+- Replace aggregation with LSM-based aggregator which has lower memory footprint {pull}11117[11117]
 - Add `service.language.name` to service destination metrics {pull}11117[11117]
 - Modify per-service transaction groups limit to consider more than service.name; Add per-service service destination groups limit and per-service service transaction groups limit {pull}11117[11117]