--- a/docs/data-model.asciidoc
+++ b/docs/data-model.asciidoc
@@ -541,12 +541,19 @@ there are limits on the number of unique groups tracked at any given time.
 
 Note that all the below limits may change in the future with further improvements.
 
-* For transaction metrics, the limits are 1000 services per GB of APM Server, and 5000 transaction
-groups per GB of APM Server. Additionally, each service may only consume up to 10% of the transaction groups,
+* For all the following metrics, they share a limit of 1000 services per GB of APM Server.
+** For transaction metrics, there is an additional limit of 5000 total transaction groups per GB of APM Server,
+and each service may only consume up to 10% of the transaction groups,
 which is 500 transaction groups per service per GB of APM Server.
-* For service-transaction metrics, the limit is 1000 service transaction groups per GB of APM Server.
-* For service-destination metrics, the limit is a constant of 10000 service destination groups.
-* For service-summary metrics, the limit is 1000 service summary groups per GB of APM Server.
+** For service-transaction metrics, there is an additional limit of 1000 total service transaction groups per GB of APM Server,
+and each service may only consume up to 10% of the service transaction groups,
+which is 100 service transaction groups per service per GB of APM Server.
+** For service-destination metrics, there is an additional limit of a constant 10000 total service destination groups,
+and each service may only consume up to 10% of the service destination groups,
+which is 1000 service destination groups per service.
+** For service-summary metrics, there is no additional limit.
+
+In the above, a service is defined as a combination of `service.name`, `service.environment`, `service.language.name` and `agent.name`.
 
 [float]
 ===== Overflows