--- a/go.mod
+++ b/go.mod
@@ -7,7 +7,7 @@ require (
 	github.com/cespare/xxhash/v2 v2.2.0
 	github.com/dgraph-io/badger/v2 v2.2007.3-0.20201012072640-f5a7e0a1c83b
 	github.com/dustin/go-humanize v1.0.1
-	github.com/elastic/apm-aggregation v0.0.0-20230630084251-60936ed32090
+	github.com/elastic/apm-aggregation v0.0.0-20230706092342-46ef2cba41a8
 	github.com/elastic/apm-data v0.1.1-0.20230628080651-9f67b9cdd993
 	github.com/elastic/beats/v7 v7.0.0-alpha2.0.20230629140819-59447f67558e
 	github.com/elastic/elastic-agent-client/v7 v7.1.2