--- a/internal/beater/config/aggregation.go
+++ b/internal/beater/config/aggregation.go
@@ -18,15 +18,12 @@
 package config
 
 const (
-	defaultTransactionAggregationHDRHistogramSignificantFigures = 2
-
 	defaultServiceDestinationAggregationMaxGroups = 10000
-
-	defaultServiceTxAggregationHDRHistogramSignificantFigures = 2
 )
 
 // AggregationConfig holds configuration related to various metrics aggregations.
 type AggregationConfig struct {
+	MaxServices         int                                 `config:"max_services"` // if <= 0 then will be set based on memory limits
 	Transactions        TransactionAggregationConfig        `config:"transactions"`
 	ServiceDestinations ServiceDestinationAggregationConfig `config:"service_destinations"`
 	ServiceTransactions ServiceTransactionAggregationConfig `config:"service_transactions"`
@@ -34,9 +31,7 @@ type AggregationConfig struct {
 
 // TransactionAggregationConfig holds configuration related to transaction metrics aggregation.
 type TransactionAggregationConfig struct {
-	MaxTransactionGroups           int `config:"max_groups"`   // if <= 0 then will be set based on memory limits
-	MaxServices                    int `config:"max_services"` // if <= 0 then will be set based on memory limits
-	HDRHistogramSignificantFigures int `config:"hdrhistogram_significant_figures" validate:"min=1, max=5"`
+	MaxGroups int `config:"max_groups"` // if <= 0 then will be set based on memory limits
 }
 
 // ServiceDestinationAggregationConfig holds configuration related to span metrics aggregation for service maps.
@@ -46,20 +41,13 @@ type ServiceDestinationAggregationConfig struct {
 
 // ServiceTransactionAggregationConfig holds configuration related to service transaction metrics aggregation.
 type ServiceTransactionAggregationConfig struct {
-	MaxGroups                      int `config:"max_groups"` // if <= 0 then will be set based on memory limits
-	HDRHistogramSignificantFigures int `config:"hdrhistogram_significant_figures" validate:"min=1, max=5"`
+	MaxGroups int `config:"max_groups"` // if <= 0 then will be set based on memory limits
 }
 
 func defaultAggregationConfig() AggregationConfig {
 	return AggregationConfig{
-		Transactions: TransactionAggregationConfig{
-			HDRHistogramSignificantFigures: defaultTransactionAggregationHDRHistogramSignificantFigures,
-		},
 		ServiceDestinations: ServiceDestinationAggregationConfig{
 			MaxGroups: defaultServiceDestinationAggregationMaxGroups,
-		},
-		ServiceTransactions: ServiceTransactionAggregationConfig{
-			HDRHistogramSignificantFigures: defaultServiceTxAggregationHDRHistogramSignificantFigures,
 		},
 	}
 }