--- a/systemtest/go.mod
+++ b/systemtest/go.mod
@@ -1,6 +1,6 @@
 module github.com/elastic/apm-server/systemtest
 
-go 1.19
+go 1.20
 
 require (
 	github.com/docker/docker v23.0.3+incompatible