… NOTICE.txt
… dev_docs/trace_metrics.md
… docs/data/elasticsearch/service_destination_metric.json
… go.sum
… internal/beater/config/aggregation_test.go
… internal/beater/config/config_test.go
… systemtest/aggregation_test.go
… systemtest/approvals/TestServiceDestinationAggregation.approved.json
… systemtest/approvals/TestServiceSummaryMetricsAggregationOverflow.approved.json
… systemtest/approvals/TestTransactionDroppedSpansStatsMetrics.approved.json
… tools/notice/overrides.json
… tools/notice/rules.json
… x-pack/apm-server/aggregation/baseaggregator/aggregator.go
… x-pack/apm-server/aggregation/baseaggregator/space.go
… x-pack/apm-server/aggregation/interval/string.go
… x-pack/apm-server/aggregation/labels/labels.go
… x-pack/apm-server/aggregation/labels/labels_test.go
… x-pack/apm-server/aggregation/servicesummarymetrics/aggregator.go
… x-pack/apm-server/aggregation/servicesummarymetrics/aggregator_test.go
… x-pack/apm-server/aggregation/servicetxmetrics/aggregator.go
… x-pack/apm-server/aggregation/servicetxmetrics/aggregator_test.go
… x-pack/apm-server/aggregation/spanmetrics/aggregator.go
… x-pack/apm-server/aggregation/spanmetrics/aggregator_test.go
… x-pack/apm-server/aggregation/txmetrics/aggregator.go
… x-pack/apm-server/aggregation/txmetrics/aggregator_test.go
… x-pack/apm-server/main.go
… x-pack/apm-server/main_test.go