@carsonip 2023-07-10
## Progress update

I've been working on understanding and optimizing the performance of LSM PoC. Performance improvements are WIP in apm-aggregation repo. But here's a glimpse of how LSM PoC fares against main:
```
                           │ /home/carson/projects/apm-server/apmbench-out/bench-main-2-merge.txt │ /home/carson/projects/apm-server/apmbench-out/bench-after-alloc-after-vtproto-2-merge.txt │
                           │                                sec/op                                │                              sec/op                                vs base                │
1000TransactionsSerial-512                                                           13.45m ± ∞ ¹                                                        21.76m ± ∞ ¹       ~ (p=0.333 n=2) ²
AgentAll-512                                                                         393.2m ± ∞ ¹                                                        418.0m ± ∞ ¹       ~ (p=0.333 n=2) ²
AgentGo-512                                                                          104.2m ± ∞ ¹                                                        103.7m ± ∞ ¹       ~ (p=0.667 n=2) ²
AgentNodeJS-512                                                                      51.13m ± ∞ ¹                                                        50.23m ± ∞ ¹       ~ (p=1.000 n=2) ²
AgentPython-512                                                                      162.8m ± ∞ ¹                                                        139.4m ± ∞ ¹       ~ (p=0.333 n=2) ²
AgentRuby-512                                                                        89.96m ± ∞ ¹                                                        88.98m ± ∞ ¹       ~ (p=0.667 n=2) ²
geomean                                                                              86.29m                                                              91.51m        +6.05%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                           │ /home/carson/projects/apm-server/apmbench-out/bench-main-2-merge.txt │ /home/carson/projects/apm-server/apmbench-out/bench-after-alloc-after-vtproto-2-merge.txt │
                           │                              events/sec                              │                            events/sec                              vs base                │
1000TransactionsSerial-512                                                           74.44k ± ∞ ¹                                                        45.95k ± ∞ ¹       ~ (p=0.333 n=2) ²
AgentAll-512                                                                         45.84k ± ∞ ¹                                                        43.21k ± ∞ ¹       ~ (p=0.333 n=2) ²
AgentGo-512                                                                          50.22k ± ∞ ¹                                                        50.50k ± ∞ ¹       ~ (p=0.667 n=2) ²
AgentNodeJS-512                                                                      39.47k ± ∞ ¹                                                        40.19k ± ∞ ¹       ~ (p=1.000 n=2) ²
AgentPython-512                                                                      43.19k ± ∞ ¹                                                        50.38k ± ∞ ¹       ~ (p=0.333 n=2) ²
AgentRuby-512                                                                        41.73k ± ∞ ¹                                                        42.21k ± ∞ ¹       ~ (p=0.667 n=2) ²
geomean                                                                              47.97k                                                              45.24k        -5.70%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                           │ /home/carson/projects/apm-server/apmbench-out/bench-main-2-merge.txt │ /home/carson/projects/apm-server/apmbench-out/bench-after-alloc-after-vtproto-2-merge.txt │
                           │                                 B/op                                 │                              B/op                                vs base                  │
1000TransactionsSerial-512                                                          5.667Mi ± ∞ ¹                                                    20.012Mi ± ∞ ¹         ~ (p=0.333 n=2) ²
AgentAll-512                                                                        251.7Mi ± ∞ ¹                                                     492.3Mi ± ∞ ¹         ~ (p=0.333 n=2) ²
AgentGo-512                                                                         63.21Mi ± ∞ ¹                                                    120.73Mi ± ∞ ¹         ~ (p=0.333 n=2) ²
AgentNodeJS-512                                                                     30.88Mi ± ∞ ¹                                                     59.81Mi ± ∞ ¹         ~ (p=0.333 n=2) ²
AgentPython-512                                                                     101.3Mi ± ∞ ¹                                                     200.2Mi ± ∞ ¹         ~ (p=0.333 n=2) ²
AgentRuby-512                                                                       59.99Mi ± ∞ ¹                                                    107.15Mi ± ∞ ¹         ~ (p=0.333 n=2) ²
geomean                                                                             50.67Mi                                                           107.3Mi        +111.75%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                           │ /home/carson/projects/apm-server/apmbench-out/bench-main-2-merge.txt │ /home/carson/projects/apm-server/apmbench-out/bench-after-alloc-after-vtproto-2-merge.txt │
                           │                              allocs/op                               │                            allocs/op                              vs base                 │
1000TransactionsSerial-512                                                           73.25k ± ∞ ¹                                                      162.97k ± ∞ ¹        ~ (p=0.333 n=2) ²
AgentAll-512                                                                         3.964M ± ∞ ¹                                                       5.090M ± ∞ ¹        ~ (p=0.333 n=2) ²
AgentGo-512                                                                          854.1k ± ∞ ¹                                                      1151.5k ± ∞ ¹        ~ (p=0.333 n=2) ²
AgentNodeJS-512                                                                      515.0k ± ∞ ¹                                                       659.8k ± ∞ ¹        ~ (p=0.333 n=2) ²
AgentPython-512                                                                      1.700M ± ∞ ¹                                                       2.103M ± ∞ ¹        ~ (p=0.333 n=2) ²
AgentRuby-512                                                                        940.8k ± ∞ ¹                                                      1182.0k ± ∞ ¹        ~ (p=0.333 n=2) ²
geomean                                                                              767.4k                                                             1.078M        +40.42%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

The numbers are quite unstable since I ran the benchmarks on my laptop. But other than 1000TransactionsSerial (similar to 1000Transactions but runs in serial instead of in parallel), all the agent benchmarks seem to be fairly similar.