@carsonip 2023-08-31
@kruskall do you mind leaving a note about where we did not return vt objects, or alternatively create an issue to do so? I'm very interested to know where we missed it.